+++ packages/client/src/components/form/checkbox.vue
@@ -1,6 +1,6 @@
 <template>
 <div
-	class="ziffeomt"
+	class="ziffeoms"
 	:class="{ disabled, checked }"
 >
 	<input
@@ -9,8 +9,8 @@
 		:disabled="disabled"
 		@keydown.enter="toggle"
 	>
-	<span ref="button" v-tooltip="checked ? $ts.itsOn : $ts.itsOff" class="button" @click.prevent="toggle">
-		<div class="knob"></div>
+	<span ref="button" v-adaptive-border v-tooltip="checked ? $ts.itsOn : $ts.itsOff" class="button" @click.prevent="toggle">
+		<i class="check fas fa-check"></i>
 	</span>
 	<span class="label">
 		<!-- TODO: 無名slotの方は廃止 -->
@@ -23,6 +23,7 @@
 <script lang="ts" setup>
 import { toRefs, Ref } from 'vue';
 import * as os from '@/os';
+import Ripple from '@/components/ripple.vue';
 
 const props = defineProps<{
 	modelValue: boolean | Ref<boolean>;
@@ -40,13 +41,16 @@ const toggle = () => {
 	emit('update:modelValue', !checked.value);
 
 	if (!checked.value) {
-
+		const rect = button.getBoundingClientRect();
+		const x = rect.left + (button.offsetWidth / 2);
+		const y = rect.top + (button.offsetHeight / 2);
+		os.popup(Ripple, { x, y, particle: false }, {}, 'end');
 	}
 };
 </script>
 
 <style lang="scss" scoped>
-.ziffeomt {
+.ziffeoms {
 	position: relative;
 	display: flex;
 	transition: all 0.2s ease;
@@ -69,24 +73,21 @@ const toggle = () => {
 		flex-shrink: 0;
 		margin: 0;
 		box-sizing: border-box;
-		width: 32px;
+		width: 23px;
 		height: 23px;
 		outline: none;
-		background: var(--swutchOffBg);
-		background-clip: content-box;
-		border: solid 1px var(--swutchOffBg);
-		border-radius: 999px;
+		background: var(--panel);
+		border: solid 1px var(--panel);
+		border-radius: 4px;
 		cursor: pointer;
 		transition: inherit;
 
-		> .knob {
-			position: absolute;
-			top: 3px;
-			left: 3px;
-			width: 15px;
-			height: 15px;
-			background: var(--swutchOffFg);
-			border-radius: 999px;
+		> .check {
+			margin: auto;
+			opacity: 0;
+			color: var(--fgOnAccent);
+			font-size: 13px;
+			transform: scale(0.5);
 			transition: all 0.2s ease;
 		}
 	}
@@ -129,12 +130,12 @@ const toggle = () => {
 
 	&.checked {
 		> .button {
-			background-color: var(--swutchOnBg) !important;
-			border-color: var(--swutchOnBg) !important;
+			background-color: var(--accent) !important;
+			border-color: var(--accent) !important;
 
-			> .knob {
-				left: 12px;
-				background: var(--swutchOnFg);
+			> .check {
+				opacity: 1;
+				transform: scale(1);
 			}
 		}
 	}