+++ packages/client/src/components/user-card-mini.vue
@@ -1,11 +1,11 @@
 <template>
-<div :class="[$style.root, { yellow: instance.isNotResponding, red: instance.isBlocked, gray: instance.isSuspended }]">
-	<img v-if="instance.iconUrl" class="icon" :src="instance.iconUrl" alt=""/>
+<div :class="[$style.root, { yellow: user.isSilenced, red: user.isSuspended, gray: false }]">
+	<MkAvatar class="avatar" :user="user" :disable-link="true" :show-indicator="true"/>
 	<div class="body">
-		<span class="host">{{ instance.host }}</span>
-		<span class="sub">{{ instance.softwareName || '?' }} {{ instance.softwareVersion }}</span>
+		<span class="name"><MkUserName class="name" :user="user"/></span>
+		<span class="sub"><span class="acct _monospace">@{{ acct(user) }}</span></span>
 	</div>
-	<MkMiniChart v-if="chart" class="chart" :src="chart.requests.received"/>
+	<MkMiniChart v-if="chartValues" class="chart" :src="chartValues"/>
 </div>
 </template>
 
@@ -13,15 +13,18 @@
 import * as misskey from 'misskey-js';
 import MkMiniChart from '@/components/mini-chart.vue';
 import * as os from '@/os';
+import { acct } from '@/filters/user';
 
 const props = defineProps<{
-	instance: misskey.entities.Instance;
+	user: misskey.entities.User;
 }>();
 
-const chart = $ref(null);
+let chartValues = $ref<number[] | null>(null);
 
-os.api('charts/instance', { host: props.instance.host, limit: 16, span: 'hour' }).then(res => {
-	chart = res;
+os.apiGet('charts/user/notes', { userId: props.user.id, limit: 16 + 1, span: 'day' }).then(res => {
+	// 今日のぶんの値はまだ途中の値であり、それも含めると大抵の場合前日よりも下降しているようなグラフになってしまうため今日は弾く
+	res.inc.splice(0, 1);
+	chartValues = res.inc;
 });
 </script>
 
@@ -36,13 +39,11 @@ os.api('charts/instance', { host: props.instance.host, limit: 16, span: 'hour' }
 	background: var(--panel);
 	border-radius: 8px;
 
-	> :global(.icon) {
+	> :global(.avatar) {
 		display: block;
 		width: ($bodyTitleHieght + $bodyInfoHieght);
 		height: ($bodyTitleHieght + $bodyInfoHieght);
-		object-fit: cover;
-		border-radius: 4px;
-		margin-right: 8px;
+		margin-right: 12px;
 	}
 
 	> :global(.body) {
@@ -52,7 +53,7 @@ os.api('charts/instance', { host: props.instance.host, limit: 16, span: 'hour' }
 		color: var(--fg);
 		padding-right: 8px;
 
-		> :global(.host) {
+		> :global(.name) {
 			display: block;
 			width: 100%;
 			white-space: nowrap;
@@ -62,7 +63,9 @@ os.api('charts/instance', { host: props.instance.host, limit: 16, span: 'hour' }
 		}
 
 		> :global(.sub) {
-			font-size: 75%;
+			display: block;
+			width: 100%;
+			font-size: 95%;
 			opacity: 0.7;
 			line-height: $bodyInfoHieght;
 			white-space: nowrap;